requirements.txt
django
gunicorn
django-heroku
django-rest-swagger
djangorestframework
turicreate
